all tests OK

+++ MovingAreasExtraction/utils.py
@@ -1,7 +1,7 @@
 import os
 import shutil
 
-def checkDir(path):
+def checkDir(path, rmExist = True):
     """
     Check if directory already exists and create it
     
@@ -9,8 +9,10 @@ def checkDir(path):
     """
 
     #Check if the path exists
-    if os.path.exists(path):
+    if os.path.exists(path) and rmExist:
         #Remove the path if it exists
         shutil.rmtree(path)    
+    elif os.path.exists(path) and not rmExist:
+        return
     #Recreate it empty
     os.mkdir(path)
